revert osqp highs roles
