+++ Project.toml
@@ -8,12 +8,12 @@ Accessors = "7d9f7c33-5ae7-4f3b-8dc6-eff91059b697"
 Dictionaries = "85a47980-9c8c-11e8-2b9f-f7ca1fa99fb4"
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
 ElasticArrays = "fdbdab4c-e67f-52f5-8c3f-e7b388dad3d4"
+HiGHS = "87dc4568-4c63-4d18-b0c0-bb2238e4078b"
 JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
 LazyArrays = "5078a376-72f3-5289-bfd5-ec5146d43c02"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
 NLopt = "76087f3c-5699-56af-9a33-bf431cd00edd"
-OSQP = "ab2f91bb-94b4-55e3-9ba0-7f65df51de79"
 OffsetArrays = "6fe1bfb0-de20-5000-8ca7-80f57d26f881"
 PaddedViews = "5432bcbf-9aad-5242-b902-cca2824c8663"
 Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
@@ -29,12 +29,12 @@ UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
 [weakdeps]
 ConcurrentUtils = "3df5f688-6c4c-4767-8685-17f5ad261477"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
-HiGHS = "87dc4568-4c63-4d18-b0c0-bb2238e4078b"
+OSQP = "ab2f91bb-94b4-55e3-9ba0-7f65df51de79"
 
 [extensions]
 ConcurrentRWLockExt = "ConcurrentUtils"
 ForwardDiffBackendExt = "ForwardDiff"
-HiGHSStepsExt = "HiGHS"
+OSQPStepsExt = "OSQP"
 
 [compat]
 julia = "1.6"
@@ -49,4 +49,4 @@ SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["SafeTestsets", "Test", "ConcurrentUtils", "ForwardDiff", "Random", "LinearAlgebra", "HiGHS"]
+test = ["SafeTestsets", "Test", "ConcurrentUtils", "ForwardDiff", "Random", "LinearAlgebra", "OSQP"]
